src/requirements.txt
streamlit==1.31.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-experimental==0.3.0
langchain-core==0.3.0
sqlalchemy==2.0.27
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
reportlab==4.0.9
kaleido==0.2.1
Pillow==10.2.0

